correctif

+++ backend/controllers/sauce.js
@@ -1,5 +1,6 @@
 const Sauce = require('../models/sauce');
 const fs = require('fs');
+const sauce = require('../models/sauce');
 
 exports.createSauce = (req, res, next) => {
     const sauceObject = JSON.parse(req.body.sauce);
@@ -18,11 +19,17 @@ exports.createSauce = (req, res, next) => {
 };
 
 exports.modifySauce = (req, res, next) => {
+
     const sauceObject = req.file ? 
         { 
             ...JSON.parse(req.body.sauce), 
             imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
         } : { ...req.body };
+        
+    if (sauceObject.userId != req.body.decodedId){
+        return res.status(403).json({error : "403: unauthorized request"})
+    }
+
     Sauce.updateOne({  _id: req.params.id }, { ...sauceObject, _id: req.params.id})
         .then(() => res.status(200).json({message: 'Sauce mise à jour'}))
         .catch(error => res.status(400).json({ error }));
@@ -30,7 +37,11 @@ exports.modifySauce = (req, res, next) => {
 
 exports.deleteSauce = (req, res, next) => {
     Sauce.findOne({ _id: req.params.id})
-        .then(sauce => {
+        .then(sauce => {  
+            if (sauce.userId != req.body.decodedId){
+                return res.status(403).json({error :"utilisateur non autorisé"});
+            }
+
             const filename = sauce.imageUrl.split('/images/')[1];
             fs.unlink(`images/${filename}`, () => {
                 Sauce.deleteOne({ _id: req.params.id})
